move check_header and Verification error to main module
