+++ crates/sp-consensus-subspace/src/verification.rs
@@ -16,9 +16,7 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 //! Verification for Subspace headers.
-use crate::digests::{CompatibleDigestItem, PreDigest};
-use crate::{find_pre_digest, FarmerPublicKey, FarmerSignature, RANDOMNESS_CONTEXT};
-use codec::Decode;
+use crate::{FarmerPublicKey, FarmerSignature, VerificationError, RANDOMNESS_CONTEXT};
 use core::mem;
 use schnorrkel::context::SigningContext;
 use schnorrkel::vrf::VRFOutput;
@@ -26,7 +24,6 @@ use schnorrkel::{PublicKey, Signature, SignatureResult};
 use sp_api::HeaderT;
 use sp_arithmetic::traits::SaturatedConversion;
 use sp_consensus_slots::Slot;
-use sp_runtime::DigestItem;
 use subspace_archiving::archiver;
 use subspace_core_primitives::{
     crypto, PieceIndex, PieceIndexHash, Randomness, Salt, Sha256Hash, Solution, Tag, TagSignature,
@@ -36,160 +33,6 @@ use subspace_solving::{
     create_tag_signature_transcript, derive_global_challenge, derive_target, is_tag_valid,
     verify_local_challenge, verify_tag_signature, SubspaceCodec,
 };
-
-/// Errors encountered by the Subspace authorship task.
-#[derive(Debug, Eq, PartialEq)]
-#[cfg_attr(feature = "thiserror", derive(thiserror::Error))]
-pub enum VerificationError<Header: HeaderT> {
-    /// No Subspace pre-runtime digest found
-    #[cfg_attr(feature = "thiserror", error("No Subspace pre-runtime digest found"))]
-    NoPreRuntimeDigest,
-    /// Header has a bad seal
-    #[cfg_attr(feature = "thiserror", error("Header {0:?} has a bad seal"))]
-    HeaderBadSeal(Header::Hash),
-    /// Header is unsealed
-    #[cfg_attr(feature = "thiserror", error("Header {0:?} is unsealed"))]
-    HeaderUnsealed(Header::Hash),
-    /// Bad reward signature
-    #[cfg_attr(feature = "thiserror", error("Bad reward signature on {0:?}"))]
-    BadRewardSignature(Header::Hash),
-    /// Bad solution signature
-    #[cfg_attr(
-        feature = "thiserror",
-        error("Bad solution signature on slot {0:?}: {1:?}")
-    )]
-    BadSolutionSignature(Slot, schnorrkel::SignatureError),
-    /// Bad local challenge
-    #[cfg_attr(
-        feature = "thiserror",
-        error("Local challenge is invalid for slot {0}: {1}")
-    )]
-    BadLocalChallenge(Slot, schnorrkel::SignatureError),
-    /// Solution is outside of solution range
-    #[cfg_attr(
-        feature = "thiserror",
-        error("Solution is outside of solution range for slot {0}")
-    )]
-    OutsideOfSolutionRange(Slot),
-    /// Solution is outside of max plot size
-    #[cfg_attr(
-        feature = "thiserror",
-        error("Solution is outside of max plot size {0}")
-    )]
-    OutsideOfMaxPlot(Slot),
-    /// Invalid encoding of a piece
-    #[cfg_attr(feature = "thiserror", error("Invalid encoding for slot {0}"))]
-    InvalidEncoding(Slot),
-    /// Invalid tag for salt
-    #[cfg_attr(feature = "thiserror", error("Invalid tag for salt for slot {0}"))]
-    InvalidTag(Slot),
-}
-
-/// A header which has been checked
-pub enum CheckedHeader<H, S> {
-    /// A header which has slot in the future. this is the full header (not stripped)
-    /// and the slot in which it should be processed.
-    Deferred(H, Slot),
-    /// A header which is fully checked, including signature. This is the pre-header
-    /// accompanied by the seal components.
-    ///
-    /// Includes the digest item that encoded the seal.
-    Checked(H, S),
-}
-
-/// Subspace verification parameters
-pub struct VerificationParams<'a, Header>
-where
-    Header: HeaderT + 'a,
-{
-    /// The header being verified.
-    pub header: Header,
-    /// The slot number of the current time.
-    pub slot_now: Slot,
-    /// Parameters for solution verification
-    pub verify_solution_params: VerifySolutionParams<'a>,
-    /// Signing context for reward signature
-    pub reward_signing_context: &'a SigningContext,
-}
-
-/// Information from verified header
-pub struct VerifiedHeaderInfo<RewardAddress> {
-    /// Pre-digest
-    pub pre_digest: PreDigest<FarmerPublicKey, RewardAddress>,
-    /// Seal (signature)
-    pub seal: DigestItem,
-}
-
-/// Check a header has been signed correctly and whether solution is correct. If the slot is too far
-/// in the future, an error will be returned. If successful, returns the pre-header and the digest
-/// item containing the seal.
-///
-/// The seal must be the last digest. Otherwise, the whole header is considered unsigned. This is
-/// required for security and must not be changed.
-///
-/// This digest item will always return `Some` when used with `as_subspace_pre_digest`.
-///
-/// `pre_digest` argument is optional in case it is available to avoid doing the work of extracting
-/// it from the header twice.
-pub fn check_header<Header, RewardAddress>(
-    params: VerificationParams<Header>,
-    pre_digest: Option<PreDigest<FarmerPublicKey, RewardAddress>>,
-) -> Result<CheckedHeader<Header, VerifiedHeaderInfo<RewardAddress>>, VerificationError<Header>>
-where
-    Header: HeaderT,
-    RewardAddress: Decode,
-{
-    let VerificationParams {
-        mut header,
-        slot_now,
-        verify_solution_params,
-        reward_signing_context,
-    } = params;
-
-    let pre_digest = match pre_digest {
-        Some(pre_digest) => pre_digest,
-        None => find_pre_digest::<Header, RewardAddress>(&header)
-            .ok_or(VerificationError::NoPreRuntimeDigest)?,
-    };
-    let slot = pre_digest.slot;
-
-    let seal = header
-        .digest_mut()
-        .pop()
-        .ok_or_else(|| VerificationError::HeaderUnsealed(header.hash()))?;
-
-    let signature = seal
-        .as_subspace_seal()
-        .ok_or_else(|| VerificationError::HeaderBadSeal(header.hash()))?;
-
-    // The pre-hash of the header doesn't include the seal and that's what we sign
-    let pre_hash = header.hash();
-
-    if pre_digest.slot > slot_now {
-        header.digest_mut().push(seal);
-        return Ok(CheckedHeader::Deferred(header, pre_digest.slot));
-    }
-
-    // Verify that block is signed properly
-    if check_reward_signature(
-        pre_hash.as_ref(),
-        &signature,
-        &pre_digest.solution.public_key,
-        reward_signing_context,
-    )
-    .is_err()
-    {
-        return Err(VerificationError::BadRewardSignature(pre_hash));
-    }
-
-    // Verify that solution is valid
-    verify_solution(&pre_digest.solution, slot, verify_solution_params)?;
-
-    Ok(CheckedHeader::Checked(
-        header,
-        VerifiedHeaderInfo { pre_digest, seal },
-    ))
-}
 
 /// Check the reward signature validity.
 pub fn check_reward_signature(
